+++ src/components/calendar/CalendarDays/style.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+import COLOR from "../../../constants/color";
+
+const CalendarDate = styled.div`
+  display: flex;
+  justify-content: space-around;
+  margin-bottom: 0.7rem;
+  color: ${COLOR.GRAY3};
+  font-size: 0.9rem;
+`;
+
+export { CalendarDate };
