+++ src/routes/Schedule/Calendar/index.tsx
@@ -1,95 +1,9 @@
 import { useState } from "react";
-import { format, addMonths, subMonths, getDay, isSameDay } from "date-fns";
-import { startOfMonth, endOfMonth, startOfWeek, endOfWeek } from "date-fns";
-import { addDays } from "date-fns";
-import {
-  CalendarCell,
-  CalendarCells,
-  CalendarContainer,
-  CalendarDate,
-  CalendarHeader,
-  CalendarHeaderArrow,
-  CalendarHeaderLeft,
-  CalendarHeaderRight,
-  CalendarToday,
-} from "./style";
-import FONT from "../../../constants/fonts";
-
-type RenderHeaderProps = {
-  currentMonth: Date;
-  prevMonth: () => void;
-  nextMonth: () => void;
-  today: () => void;
-};
-
-const RenderHeader = ({
-  currentMonth,
-  prevMonth,
-  nextMonth,
-  today,
-}: RenderHeaderProps) => {
-  return (
-    <CalendarHeader>
-      <CalendarHeaderLeft>
-        <span style={FONT.BODY1}>{format(currentMonth, "yyyy")}년</span>
-        <span style={FONT.BODY1}>{format(currentMonth, "M")}월</span>
-      </CalendarHeaderLeft>
-      <CalendarHeaderRight>
-        <CalendarHeaderArrow style={FONT.BODY1} onClick={prevMonth}>
-          &lt;
-        </CalendarHeaderArrow>
-        <CalendarToday style={FONT.BODY1} onClick={today}>
-          오늘
-        </CalendarToday>
-        <CalendarHeaderArrow style={FONT.BODY1} onClick={nextMonth}>
-          &gt;
-        </CalendarHeaderArrow>
-      </CalendarHeaderRight>
-    </CalendarHeader>
-  );
-};
-
-const RenderDays = () => {
-  const days = [];
-  const date = ["일", "월", "화", "수", "목", "금", "토"];
-  for (let i = 0; i < 7; i++) {
-    days.push(<div key={i}>{date[i]}</div>);
-  }
-  return <CalendarDate>{days}</CalendarDate>;
-};
-
-type RenderCellsProps = {
-  currentMonth: Date;
-};
-
-const RenderCells = ({ currentMonth }: RenderCellsProps) => {
-  const monthStart = startOfMonth(currentMonth);
-  const monthEnd = endOfMonth(monthStart);
-  const startDate = startOfWeek(monthStart);
-  const endDate = endOfWeek(monthEnd);
-
-  const rows = [];
-  let days = [];
-  let day = startDate;
-
-  while (day <= endDate) {
-    for (let i = 0; i < 7; i++) {
-      days.push(
-        <CalendarCells
-          key={i}
-          isWeekend={getDay(day) === 0 || getDay(day) === 6}
-          isToday={isSameDay(day, new Date())}
-        >
-          {format(day, "d")}
-        </CalendarCells>
-      );
-      day = addDays(day, 1);
-    }
-    rows.push(<CalendarCell key={rows.length}>{days}</CalendarCell>);
-    days = [];
-  }
-  return <div>{rows}</div>;
-};
+import { addMonths, subMonths } from "date-fns";
+import { CalendarContainer } from "./style";
+import CalendarHeader from "../../../components/calendar/CalendarHeader";
+import CalendarDays from "../../../components/calendar/CalendarDays";
+import CalendarCells from "../../../components/calendar/CalendarCells";
 
 const Calendar = () => {
   const [currentMonth, setCurrentMonth] = useState(new Date());
@@ -106,14 +20,14 @@ const Calendar = () => {
 
   return (
     <CalendarContainer>
-      <RenderHeader
+      <CalendarHeader
         currentMonth={currentMonth}
         prevMonth={prevMonth}
         nextMonth={nextMonth}
         today={handleTodayClick}
       />
-      <RenderDays />
-      <RenderCells currentMonth={currentMonth} />
+      <CalendarDays />
+      <CalendarCells currentMonth={currentMonth} />
     </CalendarContainer>
   );
 };
